Rename and Relocate CommandProgressView
