+++ RoboSharp.BackupApp/MainWindow.xaml
@@ -2,7 +2,9 @@
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:av="http://schemas.microsoft.com/expression/blend/2008" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" xmlns:local="clr-namespace:RoboSharp.BackupApp" mc:Ignorable="av" x:Class="RoboSharp.BackupApp.MainWindow"
-        Title="Backup App" Height="700" Width="750" ResizeMode="CanResizeWithGrip" MinWidth="750" MinHeight="650" >
+    xmlns:views="clr-namespace:RoboSharp.BackupApp.Views"
+    xmlns:viewmodels="clr-namespace:RoboSharp.BackupApp.ViewModels"
+    Title="Backup App" Height="700" Width="750" ResizeMode="CanResizeWithGrip" MinWidth="750" MinHeight="650" >
     <Grid IsSharedSizeScope="True">
         <TabControl>
             <TabItem Header="Options" x:Name="OptionsTab">
@@ -337,12 +339,15 @@
                     </ScrollViewer>
                 </Grid>
             </TabItem>
+            <TabItem Header="Extensions.BatchCommand">
+                
+            </TabItem>
             <!-- Single-Job Tab -->
             <TabItem x:Name="SingleJobTab" Header="Single Job ( RoboCommand )" >
                 <Grid x:Name="SingleJobTabGrid" Background="#FFE5E5E5">
                     <ScrollViewer Height="Auto" Margin="0,10,5,0" VerticalAlignment="Top" MinWidth="693" Grid.ColumnSpan="4">
                         <StackPanel>
-                            <local:MultiJob_CommandProgressIndicator x:Name="SingleJobExpander_Progress" RenderTransformOrigin="0.5,0.5" Background="LightBlue"/>
+                            <views:CommandProgressView x:Name="SingleJobExpander_Progress" RenderTransformOrigin="0.5,0.5" Background="LightBlue"/>
                             <Expander x:Name="SingleJobExpander_Errors" Background="LightGreen" Header="Errors">
                                 <DataGrid x:Name="SingleJobErrorGrid" MinHeight="100" Margin="10"/>
                             </Expander>
@@ -427,8 +432,8 @@
                                     <!-- Command Progress -->
                                     <ScrollViewer Grid.Row="5" Grid.ColumnSpan="2" HorizontalAlignment="Stretch" VerticalAlignment="Top">
                                         <StackPanel Name="RoboQueueProgressStackPanel" >
-                                            <local:MultiJob_CommandProgressIndicator HorizontalAlignment="Stretch" VerticalAlignment="Top" Width="600"/>
-                                            <local:MultiJob_CommandProgressIndicator HorizontalAlignment="Stretch" VerticalAlignment="Top" Width="600"/>
+                                            <views:CommandProgressView HorizontalAlignment="Stretch" VerticalAlignment="Top" Width="600"/>
+                                            <views:CommandProgressView HorizontalAlignment="Stretch" VerticalAlignment="Top" Width="600"/>
                                         </StackPanel>
                                     </ScrollViewer>
                                 </Grid>
