+++ RoboSharp.BackupApp/Views/CommandProgressView.xaml
@@ -1,4 +1,4 @@
-﻿<Expander x:Class="RoboSharp.BackupApp.MultiJob_CommandProgressIndicator"
+﻿<Expander x:Class="RoboSharp.BackupApp.Views.CommandProgressView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
